requirements.txt
fastapi[all]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv
pyjwt
